requirements.txt
google-api-python-client
oauth2client
python-dotenv